Change how static files are named, and simplify workflow
This should avoid getting duplicates of the static images with new filenames
if additional cells (e.g. Youtube video links) are inserted during the
review process.

Additionally, in figuring out how to do this, I realized we didn't need the
RSTExporter machinery, so this simplifies the build by relieving the
dependency on pandoc.