Update W0D1_Tutorial1_Solution_218a234b.py
This exercise is meant to replicate exercise 7 but using numpy arrays in place of python lists. The solution for exercise 7 and its graph shows all 150 points are plotted. This solution plots 149 points and misses the first one. You'll see this visually in comparing the 2 graphs.

The recommended code changes increase the size of the v array by one to accommodate the last calculated update to v and slices v back to size before plotting.

The solution graph will also need to be updated.